moved upload code to StorageController. trigger indexing on upload
