+++ 02 - Web UI Template/CognitiveSearch.UI/Controllers/HomeController.cs
@@ -12,7 +12,6 @@ using Microsoft.WindowsAzure.Storage.Auth;
 using CognitiveSearch.UI.Models;
 using Newtonsoft.Json.Linq;
 using Microsoft.Azure.Search.Models;
-using System.Threading.Tasks;
 
 namespace CognitiveSearch.UI.Controllers
 {
@@ -223,31 +222,6 @@ namespace CognitiveSearch.UI.Controllers
                 });
         }
 
-        [HttpPost("upload")]
-        public async Task<IActionResult> Upload()
-        {
-            if (Request.Form.Files.Any())
-            {
-                var container = GetStorageContainer();
-
-                foreach (var formFile in Request.Form.Files)
-                {
-                    if (formFile.Length > 0)
-                    {
-                        var cloudBlockBlob = container.GetBlockBlobReference(formFile.FileName);
-                        await cloudBlockBlob.UploadFromStreamAsync(formFile.OpenReadStream());
-
-                        //_telemetryClient.TrackEvent("FileUpload", telemetryDict);
-                    }
-                }
-            }
-            
-            //await _searchClient.RunIndexer();
-
-            return new JsonResult("ok");
-        }
-
-
         public class MapCredentials
         {
             public string MapKey { get; set; }
@@ -320,16 +294,6 @@ namespace CognitiveSearch.UI.Controllers
             }
 
             return tokens;
-        }
-
-        private CloudBlobContainer GetStorageContainer()
-        {
-            string accountName = _configuration.GetSection("StorageAccountName")?.Value;
-            string accountKey = _configuration.GetSection("StorageAccountKey")?.Value;
-            var containerAddress = _configuration.GetSection("StorageContainerAddress")?.Value.ToLower();
-
-            var container = new CloudBlobContainer(new Uri(containerAddress), new StorageCredentials(accountName, accountKey));
-            return container;
         }
 
         [HttpPost]
