+++ 02 - Web UI Template/CognitiveSearch.UI/Views/Shared/_Layout.cshtml
@@ -58,7 +58,6 @@
 
     <script src="https://atlas.microsoft.com/sdk/javascript/mapcontrol/2/atlas.min.js"></script>
 
-
     <script src="https://static2.sharepointonline.com/files/fabric/office-ui-fabric-js/1.2.0/js/fabric.min.js"></script>
     <script src="https://d3js.org/d3.v4.min.js"></script>
     <script src="https://d3js.org/d3-selection-multi.v1.js"></script>
@@ -72,19 +71,23 @@
 
     @await Html.PartialAsync("_Footer")
 
-    @*		<script src="~/lib/jquery/dist/jquery.js"></script>
+    @* Begin JQuery style fiddling *@
+    @*<script src="~/lib/jquery/dist/jquery.js"></script>
 
-        <script src="https://code.jquery.com/ui/1.12.1/jquery-ui.js"></script>
+    <script src="https://code.jquery.com/ui/1.12.1/jquery-ui.js"></script>
 
-        <script type="text/javascript">
-            // Change JQueryUI plugin names to fix name collision with Bootstrap.
-            $.widget.bridge('uitooltip', $.ui.tooltip);
-            $.widget.bridge('uibutton', $.ui.button);
-        </script>
-        <script src="~/lib/bootstrap/dist/js/bootstrap.js"></script>
+    <script type="text/javascript">
+        // Change JQueryUI plugin names to fix name collision with Bootstrap.
+        $.widget.bridge('uitooltip', $.ui.tooltip);
+        $.widget.bridge('uibutton', $.ui.button);
+    </script>
+    <script src="~/lib/bootstrap/dist/js/bootstrap.js"></script>
 
-        <script src="https://unpkg.com/masonry-layout@4/dist/masonry.pkgd.min.js"></script>
-        <script src="https://unpkg.com/imagesloaded@4/imagesloaded.pkgd.min.js"></script>*@
+    <script src="https://unpkg.com/masonry-layout@4/dist/masonry.pkgd.min.js"></script>
+    <script src="https://unpkg.com/imagesloaded@4/imagesloaded.pkgd.min.js"></script>*@
+    @* End JQuery style fiddling *@
+
+
     <script src="~/js/tags.js"></script>
     <script src="~/js/details.js"></script>
     <script src="~/js/appInsights.js"></script>
