+++ 02 - Web UI Template/CognitiveSearch.UI/appsettings.json
@@ -12,6 +12,7 @@
   "SearchServiceName": "matttest2search",
   "SearchApiKey": "AF4125E7FA1F24249DDFFF09437C6080",
   "SearchIndexName": "matttest2-index",
+  "SearchIndexerName": "matttest2-indexer",
   "StorageAccountName": "matttest2storage",
   "StorageAccountKey": "clCF4QqXFaGpjwST/r2KAxyvQgREsFxfMs7fH2pqSJh7xXaJzPVi2x5cQNW3+KR+Bzwz2yufgBHFI+gCEsgqdw==",
   "StorageContainerAddress": "https://matttest2storage.blob.core.windows.net/documents",
